Move LLMContent to `types` package.
