+++ packages/board-server/package.json
@@ -185,6 +185,7 @@
   },
   "dependencies": {
     "@breadboard-ai/data-store": "0.2.3",
+    "@breadboard-ai/types": "0.1.1",
     "@google-cloud/firestore": "^7.10.0",
     "@google-cloud/secret-manager": "^5.6.0",
     "@google-labs/breadboard": "^0.27.1",
