+++ packages/data-store/package.json
@@ -152,6 +152,7 @@
     "node": ">=20.14.0"
   },
   "dependencies": {
+    "@breadboard-ai/types": "0.1.1",
     "@google-labs/breadboard": "^0.27.0",
     "idb": "^8.0.0"
   }
