+++ packages/types/src/index.ts
@@ -5,3 +5,4 @@
  */
 
 export type * from "./graph-descriptor.js";
+export type * from "./llm-content.js";
