+++ packages/types/src/llm-content.ts
@@ -1,0 +1,57 @@
+/**
+ * @license
+ * Copyright 2024 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export type FunctionCallCapabilityPart = {
+  functionCall: {
+    name: string;
+    args: object;
+  };
+};
+
+export type FunctionResponseCapabilityPart = {
+  functionResponse: {
+    name: string;
+    response: object;
+  };
+};
+
+export type TextCapabilityPart = {
+  text: string;
+};
+
+export type DataStoreHandle = string;
+
+/**
+ * Represents data that is stored by a DataStoreProvider.
+ */
+export type StoredDataCapabilityPart = {
+  storedData: {
+    handle: DataStoreHandle;
+    mimeType: string;
+  };
+};
+
+export type DataPart =
+  | InlineDataCapabilityPart
+  | StoredDataCapabilityPart
+  | FunctionCallCapabilityPart
+  | FunctionResponseCapabilityPart
+  | TextCapabilityPart;
+
+export type LLMContent = {
+  role?: string;
+  parts: DataPart[];
+};
+
+/**
+ * Represents inline data, encoded as a base64 string.
+ */
+export type InlineDataCapabilityPart = {
+  inlineData: {
+    mimeType: string;
+    data: string;
+  };
+};
